// src/app/formularios/interfaces/domicilio.interface.ts
export interface Domicilio {
    idDomicilio: string;
    paisDomicilio: string;
    provinciaDomicilio: string;
    ciudadDomicilio: string;
    direccionDomicilio: string;
    casaDomicilio: string;
    edificioDomicilio: string;
    aptoDomicilio: string;
    codigoPostal: number;
    disponibleDesde: number;
    disponibleHasta: number;
    recibePersonalmente: boolean;
  }

  